% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log_plot.R
\name{data_log_plot}
\alias{data_log_plot}
\title{Scatter plot of data points}
\usage{
data_log_plot(
  x,
  y,
  type = "logistic",
  pch = 16,
  col1 = rgb(0, 0, 0, 0.6),
  col2 = rgb(1, 0, 0, 1),
  incl_mean = TRUE
)
}
\arguments{
\item{x}{(Simulated) data denoting position of data point on the x-axis.}

\item{y}{(Simulated) data denoting position of data point on the y-axis.}

\item{type}{What shape your data should be in.}

\item{pch}{Shape of your data points in the plot.}

\item{col1}{Color of your data points in the plot.}

\item{col2}{Color of your logistic curve in the plot.}

\item{incl_mean}{Whether you want the mean of your data points displayed.}
}
\value{
A plot showing how well your data points follow the logistic function.
}
\description{
Scatter plot of data points
}
\examples{
data_log_plot(x = rnorm(10,0,1), y = rbinom(10,1,0.4))
}
