% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_participants.R
\name{rank_participants}
\alias{rank_participants}
\title{Create table that ranks participants after number of missing values}
\usage{
rank_participants(df, show = 5)
}
\arguments{
\item{df}{Data frame that has been cleaned with clean_df().}

\item{show}{How many rows the table should have.}
}
\value{
Table showing the participants with most and least missing values.
}
\description{
Create table that ranks participants after number of missing values
}
\details{
This function creates the table in the summary tab. It shows the participants
with the most and least missing values and the sum and proportion of these
missing values.
}
\examples{
\dontrun{
rank_participants(menghini_2013, 5)
}
}
