% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plotter.R
\name{run_plotter}
\alias{run_plotter}
\title{Launch shiny app}
\usage{
run_plotter()
}
\description{
Launch shiny app
}
\details{
Finding and launching the main app for the AssumptionPlotter package.
}
\examples{
run_plotter()

}
