% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_logistic.R
\name{simulate_logistic}
\alias{simulate_logistic}
\title{Simulate data for logistic regression}
\usage{
simulate_logistic(n = 100, beta0 = 0, beta1 = 1)
}
\arguments{
\item{n}{Number of observations}

\item{beta0}{Intercept}

\item{beta1}{Slope}
}
\value{
A data frame with simulated data
}
\description{
Simulate data for logistic regression
}
\examples{
simulate_logistic(n = 100, beta0 = 0, beta1 = 1)
}
