% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assumption_plot}
\alias{assumption_plot}
\title{AssumptionPlotter Plot}
\usage{
assumption_plot(
  df,
  participant,
  variables,
  expected_days,
  beeps_per_day,
  include_day = TRUE,
  include_day_line = TRUE,
  impute = "none",
  add_trend = FALSE,
  trend_type = "lm",
  theme_choice = "classic",
  palette = "none",
  palette_option = "Reds",
  text_font = "sans",
  axis_text_size = 12,
  legend_text_size = 12
)
}
\arguments{
\item{df}{Selected dataset.}

\item{participant}{Participant to be plotted.}

\item{variables}{All variables to be plotted.}

\item{expected_days}{Expected days of the study}

\item{beeps_per_day}{Expected beeps per day}

\item{include_day}{Option to include day labels.}

\item{include_day_line}{Option to include day lines.}

\item{impute}{Method for imputing NA's.}

\item{add_trend}{Include a regression line showing the trend in the data.}

\item{trend_type}{Type of trend line to be passed to geom_smooth().}

\item{theme_choice}{Decide what theme the plot should have.}

\item{palette}{Chooses either no palette or custom.}

\item{text_font}{What text font the plot should have.}

\item{axis_text_size}{Adjust the text size of the plot's axes.}

\item{legend_text_size}{Adjust the text size of the legend.}

\item{palette_options}{Palette options. Any of grDevices::hcl.pals().}
}
\value{
Interactive plot in AssumptionPlotter.
}
\description{
AssumptionPlotter Plot
}
\details{
This function creates the plots in the Plots tab of AssumptionPlotter
}
\examples{
\dontrun{
assumption_plot(
  df,
  participant,
  variables,
  expected_days,
  beeps_per_day,
  include_day = TRUE,
  include_day_line = TRUE,
  impute = c("none", "mean", "mode"),
  add_trend = FALSE,
  trend_type = c("lm", "loess"),
  theme_choice = c("classic", "minimal", "bw", "void"),
  palette = c("none", "custom"),
  palette_option = "Reds",
  text_font = c("sans", "serif", "mono"),
  axis_text_size = 12,
  legend_text_size = 12
)
}

}
