% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_datasets.R
\name{clean_df}
\alias{clean_df}
\title{Edit dataset to fit plotting}
\usage{
clean_df(
  df,
  id_col = "id",
  day_col = "day",
  exp_day,
  beep_col = "beep",
  exp_beep,
  variables
)
}
\arguments{
\item{df}{The data frame.}

\item{id_col}{Provide the name of the id column}

\item{day_col}{Name of day column.}

\item{exp_day}{Expected days of the study.}

\item{beep_col}{Name of beep column.}

\item{exp_beep}{Expected beeps per day.}

\item{variables}{A vector containing the names of all columns to be analysed.
Each variable must be within quotation marks.}
}
\value{
Data file with all possible days and beeps and with a column indicating
whether a beep was missed.
}
\description{
Edit dataset to fit plotting
}
\details{
This function edits data files (from openESM). Technically,
the openesm::get_data function should read the files. However,
because the function fails to connect to zenodo, some datasets have been
manually saved in the package. To make the plotting smooth, this function
cleans the files further for analysis. To load additional datasets, go to the
dedicated zenodo webpage from openESM, copy the link for the dataset, and
paste it in the app (in the data tab under the option "paste link"). You will
also need to include some additional information from the dedicated openESM
page. If the format is correct, the analysis should still work. Make sure to
NOT pick the raw data. You can also add your own df as long as it is a csv
and the columns have the correct names.
The function also adds a column indicating whether a beep has been missed.
}
\examples{
\dontrun{

"menghini_2023_orig" <- readr::read_tsv("https://zenodo.org/records/17347538/files/0022_menghini_ts.tsv?download=1")
names <- colnames(menghini_2023_orig)[c(9:17,22:28)]

menghini_2023 <- clean_df(
  df = menghini_2023_orig,
  id_col = "id",
  day_col = "day",
  exp_day = 3,
  beep_col = "beep",
  exp_beep = 7,
  variables = names
)
}
}
