% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_datasets.R
\name{edit_df}
\alias{edit_df}
\title{Edit dataset to fit plotting}
\usage{
edit_df(df, id_col, day_col, exp_day, beep_col, exp_beep, variables)
}
\arguments{
\item{df}{The data frame. Note that the df should be in wide format.}

\item{id_col}{Provide the index for which column the participant ID is in. If there is no participant ID (the data frame is for only one participant), write "none".}

\item{day_col}{Index number for the column containing the days.}

\item{exp_day}{Expected days of the study.}

\item{beep_col}{Index number of the column containing the beeps for the day.}

\item{exp_beep}{Expected beeps per day.}

\item{variables}{A vector containing the names of all columns to be analysed. Each variable must be within quotation marks.}
}
\value{
Creates a data file in the data folder of the package.
}
\description{
Edit dataset to fit plotting
}
\details{
This function edits the data files (from openESM). Technically, the openesm::get_data function should read the files. However, because the function fails to connect to zenodo, some datasets have been manually saved in the package. To make the plotting smooth, this function cleans the files further for analysis. Load additional datasets, go to the dedicated zenodo webpage from openESM, copy the link for the dataset, and add the dataset manually in the app. If the format is correct, the analysis should still work. Make sure to NOT pick the raw data.
}
\examples{
\dontrun{
read_openESM_data("0022_menghini_ts", "https://zenodo.org/records/17347538/files/0022_menghini_ts.tsv?download=1")
}
}
