% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_bar_chart.R
\name{pie_bar_chart}
\alias{pie_bar_chart}
\title{Pie or Bar Chart of Missing Data}
\usage{
pie_bar_chart(
  df,
  participant,
  type = "pie",
  plot_all = FALSE,
  text_font = "sans",
  axis_text_size = 12,
  legend_text_size = 12,
  theme_choice = "classic"
)
}
\arguments{
\item{df}{Selected dataset.}

\item{participant}{Participant being plotted}

\item{type}{Whether it should return a pie or bar chart.}

\item{plot_all}{Whether the function should generate a plot for everyone or only.}

\item{text_font}{Font style.}

\item{axis_text_size}{Text sixe of axes.}

\item{legend_text_size}{Text size of legend.}

\item{theme_choice}{The bar chart has the same theme as the assumption plot
the participant}
}
\value{
A pie or bar chart showing the ratio of missing values for either a
participant or the entire dataset.
}
\description{
Pie or Bar Chart of Missing Data
}
\details{
This function creates a piechart of all missing data points
}
\examples{
\dontrun{
pie_bar_chart(
  df,
  participant,
  type = c("pie", "bar"),
  plot_all = FALSE,
  text_font = c("sans", "serif", "mono"),
  axis_text_size = 12,
  legend_text_size = 12,
  theme_choice = c("classic", "minimal", "bw", "void")
  )
}


}
